requirements.txt.txt
streamlit==1.14.0
sqlite3
pandas
python-dotenv
google-generativeai
